[Home] Consolidate queries; fix test
