+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 -   Adds Live and timed auctions in home view - luc
 -   Tab bar text now aligned and centered correctly - luc
 -   Adds iPad support for Auctions view layout - luc
+-   Adds <LotsByArtistsYouFollow /> component to Sales - chris
 
 ###### Consignments
