+++ src/lib/Scenes/Home/Components/Sales/Components/LotsByFollowedArtists.tsx
@@ -8,7 +8,7 @@ import { SectionHeader } from "./SectionHeader"
 import { createPaginationContainer, graphql, QueryRenderer, QueryRendererProps } from "react-relay"
 import { get } from "lodash"
 
-const PAGE_SIZE = 2
+const PAGE_SIZE = 10
 
 interface RelayProps {
   onLoad: ({ data: array }) => void
@@ -20,8 +20,8 @@ interface RelayProps {
       }
       edges: Array<{
         node: {
+          is_biddable: boolean
           artwork: {
-            __id: string
             image: {
               aspect_ratio: number
             }
@@ -45,7 +45,7 @@ const Pagination = createPaginationContainer(
   graphql.experimental`
     fragment LotsByFollowedArtists_viewer on Viewer
       @argumentDefinitions(count: { type: "Int" }, cursor: { type: "String" }) {
-      sale_artworks(first: $count, after: $cursor, include_artworks_by_followed_artists: true)
+      sale_artworks(first: $count, after: $cursor, include_artworks_by_followed_artists: false)
         @connection(key: "LotsByFollowedArtists_sale_artworks") {
         pageInfo {
           endCursor
@@ -54,7 +54,7 @@ const Pagination = createPaginationContainer(
         edges {
           cursor
           node {
-            is_biddible
+            is_biddable
             artwork {
               ...GenericGrid_artworks
             }
@@ -88,7 +88,7 @@ export class LotsByFollowedArtists extends Component<any> {
 
 function GridContainer(props: RelayProps) {
   const artworks = get(props, "viewer.sale_artworks.edges", [])
-    .filter(({ node }) => !node.is_biddible)
+    .filter(({ node }) => node.is_biddable)
     .map(({ node }) => node.artwork)
 
   if (!artworks.length) {
