+++ src/lib/Scenes/Home/Components/Sales/Components/__stories__/Sales.Components.story.tsx
@@ -1,7 +1,7 @@
-import { storiesOf } from "@storybook/react-native"
 import React from "react"
-import { ScrollView } from "react-native"
 import { LotsByFollowedArtists } from "../LotsByFollowedArtists"
+import { ScrollView } from "react-native"
+import { storiesOf } from "@storybook/react-native"
 
 storiesOf("Home/Relay/Auctions").add("LotsByFollowedArtists", () =>
   <ScrollView>
