+++ src/lib/Scenes/Home/Components/Sales/index.tsx
@@ -1,12 +1,13 @@
 import React from "react"
+import createEnvironment from "lib/relay/createEnvironment"
 import fonts from "lib/data/fonts"
 import styled from "styled-components/native"
 import { Dimensions, FlatList, SectionList, TouchableWithoutFeedback, View } from "react-native"
 import { LotsByFollowedArtists } from "./Components/LotsByFollowedArtists"
+import { QueryRenderer, QueryRendererProps, createFragmentContainer, graphql } from "react-relay"
 import { SaleList } from "./Components/SaleList"
 import { SectionHeader } from "./Components/SectionHeader"
 import { StyleSheet, TextStyle } from "react-native"
-import { createFragmentContainer, graphql } from "react-relay"
 
 interface Props {
   sales: Array<{
@@ -64,10 +65,27 @@ class Sales extends React.Component<Props> {
   }
 }
 
+export function SalesRenderer({ render }) {
+  return (
+    <QueryRenderer
+      environment={createEnvironment()}
+      query={graphql`
+        query SalesRendererQuery {
+          sales: sales(live: true, is_auction: true) {
+            ...Sales_sales
+          }
+        }
+      `}
+      variables={{}}
+      render={render}
+    />
+  )
+}
+
 export default createFragmentContainer(Sales, {
   sales: graphql`
     fragment Sales_sales on Sale @relay(plural: true) {
-      ...SaleItem_sale
+      ...SaleListItem_sale
       live_start_at
       href
     }
