+++ src/lib/Scenes/Home/__tests__/index-tests.tsx
@@ -1,8 +1,7 @@
-import React from "react"
-
-import { shallow } from "enzyme"
-import { Tab } from "lib/Components/TabBar"
 import Home from "../index"
+import React from "react"
+import { Tab } from "lib/Components/TabBar"
+import { shallow } from "enzyme"
 
 it("has the correct number of tabs", () => {
   const wrapper = shallow(<Home />)
