+++ src/lib/Scenes/Home/index.tsx
@@ -8,11 +8,10 @@ import { Router } from "lib/utils/router"
 
 import WorksForYou from "lib/Containers/WorksForYou"
 import ForYou from "./Components/ForYou"
-import Sales from "./Components/Sales"
+import Sales, { SalesRenderer } from "./Components/Sales"
 
 import { ForYouRenderer, WorksForYouRenderer } from "lib/relay/QueryRenderers"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
-import SalesRenderer from "./Components/Sales/SalesRenderer"
 
 import DarkNavigationButton from "lib/Components/Buttons/DarkNavigationButton"
 import TabBar, { Tab } from "lib/Components/TabBar"
